summer-ospp/ModelRain/requirements.txt
mindspore==2.0.0
numpy==1.21.6
numpy==1.21.5
pandas==1.3.5
ruamel.base==1.0.0
